+++ src/AppRoutes.tsx
@@ -2,7 +2,7 @@ import { Routes, Route } from 'react-router-dom';
 import App from './App'
 import Camera from './Camera'
 import Geo from './Geo'
-import Maps from './pages/Maps/GoogleMaps'
+import GoogleMaps from './pages/Map/GoogleMaps'
 
 const AppRoutes = () => {
     return(
@@ -10,7 +10,7 @@ const AppRoutes = () => {
             <Route path='/' element={<App />} />
             <Route path='/Camera' element={<Camera />} />
             <Route path='/Geo' element={<Geo />} />
-            <Route path='/Maps' element={<Maps />} />
+            <Route path='/GoogleMaps' element={<GoogleMaps />} />
         </Routes>
     )
 }
